Fix issue where sortState will reset when already empty
The problem with always resetting this is that the change will
cause a fetch on any model that uses this. If you want to reset
the sort state as an initialization step of a controller for
instance you may inadvertantly cause a fetch on model or collection
that doesn't have correct data to make that request.